refactor: restructure explorer components
