+++ DashAI/front/src/components/notebooks/explorer/tabs/Columns.jsx
@@ -1,15 +1,14 @@
 import React, { useState } from "react";
-import PropTypes from "prop-types";
 
 import { Grid, ToggleButton, ToggleButtonGroup } from "@mui/material";
 
-import JsonDisplayer from "../../shared/JsonDisplayer";
+import JsonDisplayer from "../../../shared/JsonDisplayer";
 
 /**
- * Component that displays the parameters associated with an object.
+ * Component that displays the columns associated with an object.
  * @param {object} data object that contains all the necesary info
  */
-function Parameters({ data }) {
+function Columns({ data }) {
   const [displayMode, setDisplayMode] = useState("nested-list");
 
   return (
@@ -32,13 +31,9 @@ function Parameters({ data }) {
       </Grid>
 
       {/* JSON object display */}
-      <JsonDisplayer displayMode={displayMode} name="Parameters" data={data} />
+      <JsonDisplayer displayMode={displayMode} name="Columns" data={data} />
     </Grid>
   );
 }
 
-Parameters.propTypes = {
-  data: PropTypes.object.isRequired,
-};
-
-export default Parameters;
+export default Columns;
