+++ DashAI/front/src/components/notebooks/explorer/visualizations/ImageVisualizer.jsx
