+++ DashAI/front/src/components/notebooks/explorer/visualizations/PlotlyJsonVisualizer.jsx
