+++ DashAI/front/src/components/notebooks/explorer/visualizations/TabularVisualizer.jsx
