--- DashAI/front/src/components/notebooks/explorerVisualizations/index.jsx
@@ -1,3 +1,0 @@
-export { default as TabularVisualizer } from "./TabularVisualizer";
-export { default as PlotlyJsonVisualizer } from "./PlotlyJsonVisualizer";
-export { default as ImageVisualizer } from "./ImageVisualizer";
